(#537) Preprocessor: extract MacroExpansionEngine into a separate type
