+++ Cesium.Preprocessor/CPreprocessor.cs
@@ -1,4 +1,3 @@
-using System.Globalization;
 using System.Text;
 using Cesium.Core;
 using Cesium.Core.Warnings;
@@ -18,6 +17,8 @@ public record CPreprocessor(
     IMacroContext MacroContext,
     IWarningProcessor WarningProcessor)
 {
+    private MacroExpansionEngine _macroExpansion = new(WarningProcessor, MacroContext);
+
     public async Task<string> ProcessSource()
     {
         var buffer = new StringBuilder();
@@ -49,432 +50,6 @@ public record CPreprocessor(
         }
 
         return file.Ok;
-    }
-
-    private IEnumerable<IToken<CPreprocessorTokenType>> ExpandMacros(
-        IEnumerable<IToken<CPreprocessorTokenType>> tokens)
-    {
-        using var lexer = new TransactionalLexer(tokens, WarningProcessor);
-        while (!lexer.IsEnd)
-        {
-            var token = lexer.Consume();
-            if (token.Kind == PreprocessingToken)
-            {
-                var macroName = token.Text;
-                if (!MacroContext.TryResolveMacro(macroName, out var parameters, out var replacement))
-                {
-                    yield return token;
-                    continue;
-                }
-
-                var maybeArguments = ParseArguments(parameters, lexer);
-                if (maybeArguments is not {} arguments)
-                {
-                    // Not a macro call, just emit the token.
-                    yield return token;
-                    continue;
-                }
-
-                if (arguments.IsError)
-                    RaisePreprocessorParseError(arguments.Error);
-
-                foreach (var replaced in ExpandMacros(SubstituteMacroArguments(token, arguments.Ok, replacement)))
-                {
-                    yield return replaced;
-                }
-            }
-            else
-            {
-                yield return token;
-            }
-        }
-    }
-
-    /// <returns><c>null</c> ⇒ do not expand, non-<c>null</c> ⇒ expand if ok, throw error if not ok.</returns>
-    private ParseResult<MacroArguments>? ParseArguments(MacroParameters? parameters, TransactionalLexer lexer)
-    {
-        using var transaction = lexer.BeginTransaction();
-
-        if (parameters == null)
-        {
-            MacroArguments emptyResult = new(new(), new());
-            return transaction.End<MacroArguments>(ParseResult.Ok(emptyResult, 0));
-        }
-
-        if (lexer.IsEnd)
-        {
-            // A macro has some parameters, but we are at the end of the token stream, so no arguments is available. We
-            // should just skip expanding this macro.
-            var location = lexer.LastToken?.Location
-                           ?? new SourceLocationInfo(CompilationUnitPath, null, null);
-            transaction.End(ParseResult.Error(",", null, location, "macro arguments"));
-            return null;
-        }
-
-        if (Consume() is var leftBrace and not { Kind: LeftParen })
-        {
-            SourceLocationInfo location = leftBrace.Location;
-            transaction.End(ParseResult.Error(",", leftBrace, location, "macro arguments"));
-            return null; // no braces at all means we should just skip expanding this macro
-        }
-
-        var namedArguments = new Dictionary<string, List<IToken<CPreprocessorTokenType>>>();
-        var isFirstArgument = true;
-        foreach (var parameterToken in parameters.Parameters)
-        {
-            if (isFirstArgument)
-            {
-                isFirstArgument = false;
-            }
-            else if (Consume() is var comma and not { Text: "," or ")" })
-            {
-                SourceLocationInfo location = comma.Location;
-                return transaction.End(ParseResult.Error(",", comma, location, "macro arguments"));
-            }
-
-            var name = parameterToken.Text;
-            var argument = ParseArgument(lexer);
-            if (argument.IsError)
-                return transaction.End(argument.Error);
-
-            namedArguments[name] = argument.Ok;
-        }
-
-        var varArgs = new List<List<IToken<CPreprocessorTokenType>>>();
-        if (parameters.HasEllipsis)
-        {
-            while (Peek() is not { Kind: RightParen })
-            {
-                if (isFirstArgument)
-                {
-                    isFirstArgument = false;
-                }
-                else if (Consume() is var comma and not { Text: "," })
-                {
-                    SourceLocationInfo location = comma.Location;
-                    return transaction.End(ParseResult.Error(",", comma, location, "macro arguments"));
-                }
-
-                var varArg = ParseArgument(lexer);
-                if (varArg.IsError)
-                    return transaction.End(varArg.Error);
-
-                varArgs.Add(varArg.Ok);
-            }
-        }
-
-        if (Consume() is var token and not { Kind: RightParen })
-        {
-            SourceLocationInfo location = token.Location;
-            return transaction.End(ParseResult.Error(")", token, location, "macro arguments"));
-        }
-
-        var result = new MacroArguments(namedArguments, varArgs);
-        return transaction.End<MacroArguments>(ParseResult.Ok(result, 0));
-
-        IToken<CPreprocessorTokenType> Consume()
-        {
-            IToken<CPreprocessorTokenType> currentToken;
-            do
-            {
-                currentToken = lexer.Consume();
-            } while (currentToken is { Kind: WhiteSpace or Comment });
-
-            return currentToken;
-        }
-
-        IToken<CPreprocessorTokenType> Peek()
-        {
-            IToken<CPreprocessorTokenType> currentToken;
-            var index = 0;
-            do
-            {
-                currentToken = lexer.Peek(index++);
-            } while (currentToken is { Kind: WhiteSpace or Comment });
-
-            return currentToken;
-        }
-    }
-
-    private ParseResult<List<IToken<CPreprocessorTokenType>>> ParseArgument(TransactionalLexer lexer)
-    {
-        using var transaction = lexer.BeginTransaction();
-
-        if (lexer.IsEnd)
-            return ParseResult.Error("argument", "end of stream", 0, "macro argument");
-
-        SourceLocationInfo argumentStartLocation = lexer.Peek().Location;
-        var argument = new List<IToken<CPreprocessorTokenType>>();
-        while (!lexer.IsEnd && lexer.Peek() is not ({ Kind: RightParen } or { Text: "," }))
-        {
-            var token = lexer.Consume();
-            argument.Add(token);
-            if (token is { Kind: LeftParen })
-            {
-                var tail = ParseNestedParenthesesBlock(token.Location);
-                if (!tail.IsOk)
-                    return transaction.End(tail.Error);
-            }
-        }
-
-        if (lexer.IsEnd)
-            return transaction.End(ParseResult.Error(") or ,", null, argumentStartLocation, "macro argument"));
-
-        var processedArgument = TrimStartingWhitespace(ExpandMacros(argument)).ToList();
-        return transaction.End<List<IToken<CPreprocessorTokenType>>>(ParseResult.Ok(processedArgument, 0));
-
-        ParseResult<object?> ParseNestedParenthesesBlock(SourceLocationInfo start)
-        {
-            while (!lexer.IsEnd && lexer.Peek() is not { Kind: RightParen })
-            {
-                var token = lexer.Consume();
-                argument.Add(token);
-                if (token is { Kind: LeftParen })
-                {
-                    var tail = ParseNestedParenthesesBlock(token.Location);
-                    if (!tail.IsOk)
-                        return tail;
-                }
-            }
-
-            if (lexer.IsEnd)
-                return ParseResult.Error("terminated macro argument", null, start, "macro argument nested parentheses block");
-
-            var rightParen = lexer.Consume();
-            argument.Add(rightParen);
-            return ParseResult.Ok<object?>(null, 0);
-        }
-
-        IEnumerable<IToken<CPreprocessorTokenType>> TrimStartingWhitespace(
-            IEnumerable<IToken<CPreprocessorTokenType>> tokens) =>
-            tokens.SkipWhile(t => t is { Kind: WhiteSpace or Comment or NewLine });
-    }
-
-    /// <remarks>ISO C Standard, section 6.10.4.1 Argument substitution.</remarks>
-    private IEnumerable<IToken<CPreprocessorTokenType>> SubstituteMacroArguments(
-        IToken macroNameToken,
-        MacroArguments arguments,
-        IEnumerable<IToken<CPreprocessorTokenType>> replacement)
-    {
-        switch (macroNameToken.Text)
-        {
-            case "__FILE__":
-                yield return new Token<CPreprocessorTokenType>(
-                    macroNameToken.Range,
-                    macroNameToken.Location,
-                    "\"" + macroNameToken
-                        .Location
-                        .File?
-                        .Path
-                        .Replace("\\", "\\\\") + "\"",
-                    PreprocessingToken);
-                yield break;
-            case "__LINE__":
-            {
-                var line = macroNameToken.Location.Range.Start.Line + 1;
-                yield return new Token<CPreprocessorTokenType>(
-                    macroNameToken.Range,
-                    macroNameToken.Location,
-                    line.ToString(CultureInfo.InvariantCulture),
-                    PreprocessingToken);
-                yield break;
-            }
-        }
-
-        using var lexer = new TransactionalLexer(replacement, WarningProcessor);
-        var spaceBuffer = new List<IToken<CPreprocessorTokenType>>();
-        IEnumerable<IToken<CPreprocessorTokenType>> ClearSpaceBuffer()
-        {
-            foreach (var space in spaceBuffer)
-            {
-                yield return space;
-            }
-
-            spaceBuffer.Clear();
-        }
-
-        while (!lexer.IsEnd)
-        {
-            var token = lexer.Consume();
-            switch (token)
-            {
-                case { Kind: WhiteSpace }:
-                    spaceBuffer.Add(token);
-                    break;
-                case { Text: "#" } when PeekSignificant() is { Kind: PreprocessingToken }:
-                {
-                    var next = ConsumeSignificant();
-                    var sequence = ExpandMacros(ProcessTokenNoHash(next));
-
-                    foreach (var space in ClearSpaceBuffer())
-                    {
-                        yield return space;
-                    }
-
-                    yield return new Token<CPreprocessorTokenType>(
-                        next.Range,
-                        next.Location,
-                        Stringify(sequence),
-                        PreprocessingToken);
-                    break;
-                }
-                case { Text: "##" }:
-                {
-                    // Drop buffered spaces.
-                    spaceBuffer.Clear();
-
-                    if (PeekSignificant() is null)
-                    {
-                        throw new PreprocessorException(
-                            token.Location,
-                            "## cannot appear at the end of a macro replacement list.");
-                    }
-
-                    var next = ConsumeSignificant();
-                    var sequence = ExpandMacros(ProcessTokenNoHash(next));
-
-                    // TODO: Figure out what to do if the sequence is more than one item.
-                    yield return new Token<CPreprocessorTokenType>(
-                        next.Range,
-                        next.Location,
-                        sequence.Single().Text,
-                        PreprocessingToken);
-                    break;
-                }
-                default:
-                {
-                    foreach (var space in ClearSpaceBuffer())
-                    {
-                        yield return space;
-                    }
-
-                    var sequence = ExpandMacros(ProcessTokenNoHash(token));
-                    foreach (var item in sequence)
-                    {
-                        yield return item;
-                    }
-                    break;
-                }
-            }
-        }
-
-        foreach (var space in spaceBuffer)
-        {
-            yield return space;
-        }
-        yield break;
-
-        IToken<CPreprocessorTokenType>? PeekSignificant()
-        {
-            for (var i = 0; !lexer.IsEnd; ++i)
-            {
-                var currentToken = lexer.Peek(i);
-                if (currentToken is not { Kind: WhiteSpace or Comment or NewLine })
-                    return currentToken;
-            }
-
-            return null;
-        }
-
-        IToken<CPreprocessorTokenType> ConsumeSignificant()
-        {
-            IToken<CPreprocessorTokenType> currentToken;
-            do
-            {
-                currentToken = lexer.Consume();
-            } while (currentToken is { Kind: WhiteSpace or Comment or NewLine });
-
-            return currentToken;
-        }
-
-        IEnumerable<IToken<CPreprocessorTokenType>> ProcessTokenNoHash(IToken<CPreprocessorTokenType> token)
-        {
-            switch (token)
-            {
-                case { Kind: PreprocessingToken, Text: var text }
-                    when arguments.Named.TryGetValue(text, out var argument):
-                {
-                    // macro argument substitution
-                    foreach (var argumentToken in argument)
-                    {
-                        yield return argumentToken;
-                    }
-
-                    break;
-                }
-                // TODO: __VA_OPT__, see also rules for __VA_ARGS__ regarding the nested expansion.
-                case { Kind: PreprocessingToken, Text: "__VA_ARGS__" }:
-                {
-                    var isFirst = true;
-                    foreach (var varArg in arguments.VarArg)
-                    {
-                        // Comma before each but the first.
-                        if (isFirst) isFirst = false;
-                        else
-                        {
-                            yield return new Token<CPreprocessorTokenType>(
-                                new Range(),
-                                new Location(),
-                                ",",
-                                Separator);
-                        }
-
-                        foreach (var argToken in varArg)
-                            yield return argToken;
-                    }
-
-                    break;
-                }
-                default:
-                    yield return token;
-                    break;
-            }
-        }
-
-        string Stringify(IEnumerable<IToken<CPreprocessorTokenType>> tokens)
-        {
-            // According to the standard, and whitespace sequence gets replaced with a single space.
-            var replaced = ReplaceWhitespace();
-            var builder = new StringBuilder("\"");
-            foreach (var token in replaced)
-            {
-                foreach (var c in token.Text)
-                {
-                    builder.Append(c switch
-                    {
-                        '\\' or '\"' => "\\" + c,
-                        _ => c.ToString()
-                    });
-                }
-            }
-            builder.Append('"');
-            return builder.ToString();
-
-            IEnumerable<IToken<CPreprocessorTokenType>> ReplaceWhitespace()
-            {
-                var spaceEater = false;
-                foreach (var token in tokens)
-                {
-                    if (token is { Kind: WhiteSpace or Comment or NewLine })
-                    {
-                        if (spaceEater) continue;
-
-                        spaceEater = true;
-                        var spaceToken = token.Text == " " ? token : new Token<CPreprocessorTokenType>(
-                            token.Range,
-                            token.Location,
-                            " ",
-                            WhiteSpace);
-                        yield return spaceToken;
-                    }
-                    else
-                    {
-                        spaceEater = false;
-                        yield return token;
-                    }
-                }
-            }
-        }
     }
 
     private async IAsyncEnumerable<IToken<CPreprocessorTokenType>> ProcessGroup(IEnumerable<IGroupPart> group)
@@ -602,7 +177,7 @@ public record CPreprocessor(
             case EmptyDirective:
                 break;
             case TextLine textLine:
-                foreach (var token in ExpandMacros(textLine.Tokens ?? []))
+                foreach (var token in _macroExpansion.ExpandMacros(textLine.Tokens ?? []))
                 {
                     yield return token;
                 }
@@ -709,7 +284,7 @@ public record CPreprocessor(
         yield return new Token<CPreprocessorTokenType>(new Range(), new Location(), "\n", NewLine);
     }
 
-    private void RaisePreprocessorParseError(ParseError error)
+    internal static void RaisePreprocessorParseError(ParseError error)
     {
         var got = error.Got switch
         {
@@ -718,7 +293,7 @@ public record CPreprocessor(
             null => "<no token>"
         };
 
-        var errorMessage = new StringBuilder($"Error during preprocessing file \"{CompilationUnitPath}\", {error.Position}. Found {got}, ");
+        var errorMessage = new StringBuilder($"Error during preprocessing. Found {got}, ");
         if (error.Elements.Count == 1)
         {
             errorMessage.Append($"but expected {ExpectedString(error.Elements.Single())}");
@@ -732,15 +307,10 @@ public record CPreprocessor(
             }
         }
 
-        var location = error.Position as SourceLocationInfo ?? new SourceLocationInfo(CompilationUnitPath, null, null);
+        var location = error.Position as SourceLocationInfo ?? new SourceLocationInfo("<unknown>", null, null);
         throw new PreprocessorException(location, errorMessage.ToString());
 
         static string ExpectedString(KeyValuePair<string, ParseErrorElement> element) =>
             string.Join(", ", element.Value.Expected) + $" (rule {element.Key})";
     }
-
-    private record MacroArguments(
-        Dictionary<string, List<IToken<CPreprocessorTokenType>>> Named,
-        List<List<IToken<CPreprocessorTokenType>>> VarArg
-    );
 }
